refactor: add basic file structure for transfer component
