+++ src/Transfer/Item.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Item = ({ id, children }: { id: string; children: React.ReactNode }) => {
+  return (
+    <>
+      Item {id} {children}
+    </>
+  );
+};
+
+export { Item };
