+++ src/Transfer/ResetButton.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ResetButton = () => {
+  return <>ResetButton</>;
+};
+
+export { ResetButton };
