+++ src/Transfer/Source.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Source = ({ children }: { children: React.ReactNode }) => {
+  return <>Source {children}</>;
+};
+
+Source.Header = () => {
+  return <>Source.Header</>;
+};
+
+export { Source };
