+++ src/Transfer/TranferButton.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const TransferButton = () => {
+  return <>TransferButton</>;
+};
+
+export { TransferButton };
